# src/static/api/api.yml
openapi: 3.0.0
info:
  title: Biology Database API
  description: 'Api rest to register users in a database for genetic biological applications. Go to [Home page](http://localhost:5000/) for more information.'
  version: 1.0.0
  termsOfService: 'http://localhost:5000'
  contact:
    email: rodrjuniorsantos@gmail.com
servers:
  - url: 'http://localhost:5000'
    description: 'Primary server, e.g. Main (production) server'
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: api-token
paths:
...
  /api/entities/entity:
    post:
      summary: Create Profession Function.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      requestBody:
        description: Entity Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - documento
              properties:
                name:
                  type: string
                documento:
                  type: string
                image:
                  type: string
                  
      responses:
        '200':
          description: Entity successfully registered
        '400':
          description: Missing data
        '401':
          description: Document already exists for this user, please supply another document
        '402':
          description: There is already a registration for this user
  /api/entities/entities_all:
    get:
      summary: Get all entities Function
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      
      responses:
        '200':
          description: Returns all entities
        '400':
          description: User not found
        '401':
          description: Permission denied
  /api/entities/{entity_id}:
    get:
      summary: Gets a entity by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      parameters:
        - in: path
          name: entity_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The entity ID.

      responses:
        '200':
          description: Data this entity successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
  /api/entities/edit/{entity_id}:
    put:
      summary: Update A Entity.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      parameters:
        - in: path
          name: entity_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The Entity ID.
      requestBody:
        description: Entity Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - documento
              properties:
                name:
                  type: string
                documento:
                  type: string
                image:
                  type: string
      responses:
        '200':
          description: Entity successfully update
        '400':
          description: Entity not found
        '401':
          description: Permission denied
        '402':
          description: Missing data
        '403':
          description: Document already exists for this user, please supply another document
        '404':
          description: There is already a registration for this user
  /api/entities/delete/{entity_id}:
    delete:
      summary: Delete a entity by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Entity
      parameters:
        - in: path
          name: entity_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The entity ID.
      responses:
        '200':
          description: Entity successfully deleted
        '400':
          description: Entity not found
        '401':
          description: Permission denied
...
  /api/addresss/address:
    post:
      summary: Create Adress Function.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      requestBody:
        description: Adress Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - cidade
                - uf
                - rua
                - bairro
                - cep
              properties:
                name:
                  type: string
                cidade:
                  type: string
                uf:
                  type: string
                rua:
                  type: string
                bairro:
                  type: string
                cep:
                  type: string
                  
      responses:
        '200':
          description: Adress successfully registered
        '400':
          description: Missing data
  /api/addresss/addresss_all:
    get:
      summary: Get all address Function
      security:
        - APIKeyHeader: []
      tags:
        - Address
      
      responses:
        '200':
          description: Returns all address
        '400':
          description: User not found
        '401':
          description: Permission denied
  /api/addresss/{address_id}:
    get:
      summary: Gets a address by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      parameters:
        - in: path
          name: address_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The address ID.

      responses:
        '200':
          description: Data this address successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
  /api/addresss/edit/{address_id}:
    put:
      summary: Update A Address.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      parameters:
        - in: path
          name: address_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The Address ID.
      requestBody:
        description: Adress Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - cidade
                - uf
                - rua
                - bairro
                - cep
              properties:
                name:
                  type: string
                cidasde:
                  type: string
                uf:
                  type: string
                rua:
                  type: string
                bairro:
                  type: string
                cep:
                  type: string
      responses:
        '200':
          description: Address successfully update
        '400':
          description: Address not found
        '401':
          description: Permission denied
        '402':
          description: Missing data
  /api/addresss/delete/{address_id}:
    delete:
      summary: Delete a address by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      parameters:
        - in: path
          name: address_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The address ID.
      responses:
        '200':
          description: Address successfully deleted
        '400':
          description: Address not found
        '401':
          description: Permission denied
...